Eliminate BaseSyncTest._createCaseStubs
